Refactor durable log shipping, fix background level check when no file exists
