+++ .gitignore
@@ -186,3 +186,4 @@ UpgradeLog*.htm
 # Microsoft Fakes
 FakesAssemblies/
 *.orig
+/sample/Sample/logs/
